Moved CryptoStatic.decodeCertificate() to CryptoUtils
