+++ platform-sdk/swirlds-platform-core/src/main/java/com/swirlds/platform/crypto/CryptoStatic.java
@@ -24,10 +24,8 @@ import com.swirlds.platform.system.address.Address;
 import com.swirlds.platform.system.address.AddressBook;
 import edu.umd.cs.findbugs.annotations.NonNull;
 import edu.umd.cs.findbugs.annotations.Nullable;
-import java.io.ByteArrayInputStream;
 import java.io.FileInputStream;
 import java.io.IOException;
-import java.io.InputStream;
 import java.math.BigInteger;
 import java.nio.file.Files;
 import java.nio.file.Path;
@@ -45,7 +43,6 @@ import java.security.SignatureException;
 import java.security.UnrecoverableKeyException;
 import java.security.cert.Certificate;
 import java.security.cert.CertificateException;
-import java.security.cert.CertificateFactory;
 import java.security.cert.X509Certificate;
 import java.util.ArrayList;
 import java.util.Arrays;
@@ -221,22 +218,6 @@ public final class CryptoStatic {
                     .getCertificate(v3CertBldr.build(signerBuilder.build(caPair.getPrivate())));
         } catch (CertificateException | OperatorCreationException e) {
             throw new KeyGeneratingException("Could not generate certificate!", e);
-        }
-    }
-
-    /**
-     * Decode a X509Certificate from a byte array that was previously obtained via X509Certificate.getEncoded().
-     *
-     * @param encoded a byte array with an encoded representation of a certificate
-     * @return the certificate reconstructed from its encoded form
-     */
-    @NonNull
-    public static X509Certificate decodeCertificate(@NonNull final byte[] encoded) {
-        try (final InputStream in = new ByteArrayInputStream(encoded)) {
-            final CertificateFactory factory = CertificateFactory.getInstance("X.509");
-            return (X509Certificate) factory.generateCertificate(in);
-        } catch (CertificateException | IOException e) {
-            throw new CryptographyException(e);
         }
     }
