+++ platform-sdk/swirlds-platform-core/src/main/java/com/swirlds/platform/roster/RosterUtils.java
@@ -9,7 +9,6 @@ import com.hedera.node.internal.network.Network;
 import com.hedera.node.internal.network.NodeMetadata;
 import com.hedera.pbj.runtime.io.buffer.Bytes;
 import com.swirlds.common.RosterStateId;
-import com.swirlds.platform.crypto.CryptoStatic;
 import com.swirlds.platform.state.service.ReadableRosterStore;
 import com.swirlds.platform.state.service.WritableRosterStore;
 import com.swirlds.platform.system.address.Address;
@@ -29,6 +28,7 @@ import java.util.Objects;
 import java.util.function.Function;
 import java.util.stream.Collectors;
 import java.util.stream.IntStream;
+import org.hiero.base.crypto.CryptoUtils;
 import org.hiero.base.crypto.CryptographyException;
 import org.hiero.base.crypto.Hash;
 import org.hiero.consensus.model.node.NodeId;
@@ -75,7 +75,7 @@ public final class RosterUtils {
      */
     public static X509Certificate fetchGossipCaCertificate(@NonNull final RosterEntry entry) {
         try {
-            return CryptoStatic.decodeCertificate(entry.gossipCaCertificate().toByteArray());
+            return CryptoUtils.decodeCertificate(entry.gossipCaCertificate().toByteArray());
         } catch (final CryptographyException e) {
             return null;
         }
@@ -348,7 +348,7 @@ public final class RosterUtils {
 
         X509Certificate sigCert;
         try {
-            sigCert = CryptoStatic.decodeCertificate(entry.gossipCaCertificate().toByteArray());
+            sigCert = CryptoUtils.decodeCertificate(entry.gossipCaCertificate().toByteArray());
         } catch (final CryptographyException e) {
             // Malformed or missing gossip certificates are nullified.
             // https://github.com/hashgraph/hedera-services/issues/16648
